requirements.txt
Django==1.8.3
django-carton==1.2.0
Pillow==2.9.0
requests==2.7.0
stripe==1.25.0
wheel==0.24.0
psycopg2
dj-database-url==0.3.0
gunicorn==19.1.1
django-storages-redux
boto
